refactor(xmlrpc): extract call tests to sibling file
